% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{allocationPlot}
\alias{allocationPlot}
\title{Plot subject allocations}
\usage{
allocationPlot(
  subjectData,
  dose = Dose,
  tox = Toxicity,
  doseGrid,
  toxCategories = list(),
  toxData = NA,
  toxVar = NA,
  recommendedDose = NA
)
}
\arguments{
\item{subjectData}{The tibble containing subject level dose allocations and toxicity status}

\item{dose}{the name of the column in \code{subjectData} which defines the dose}

\item{tox}{the name of the column in \code{subjectData} which defines the toxicity status.}

\item{doseGrid}{the values to appear along the dose axis.  Can be omitted: if so, only those
values that exist in \code{dose} column of \code{subjectData} will be used.}

\item{toxCategories}{a list containing the labels of the toxicity categories that will appear
in the legend.  Can be omitted.  If omitted, sensible defaults are derived, based on the
mighest value of \code{tox} in \code{subjectData}.  Names of list elements should be "0" ... "<n>",
values are used as labels in the legend.}

\item{toxData}{the name of the tibble containing the poasterior estimates of toxicity.
See Usage notes below.}

\item{toxVar}{the name of the column in \code{toxData} containing p(Tox) [or other relevant statistic].
See Usage notes below}

\item{recommendedDose}{Optional.  The value of \code{dose} to mark as the recommended dose}
}
\description{
Plot subject allocations
}
\section{Usage notes}{

\code{toxData} and \code{toxVar} must either both be omitted or both be present.
   The \code{toxData} tibble should contain columns named \code{dose} and \code{toxVar}.
   Other columns are ignored.  It should contain one row for each value of \code{dose} in
   \code{subjectData}
}

\examples{
#Basic usage
data("oQuigleyPatientData")
allocationPlot(oQuigleyPatientData, doseGrid=1:6)
#More advanced
data("oQuigleyPosteriorShort")
toxData <- oQuigleyPosteriorShort \%>\% tidy() \%>\% augmentOQuigleyData() \%>\% createDoseSummary()
allocationPlot(oQuigleyPatientData, doseGrid=1:6, toxData=toxData, toxVar=Mean, recommendedDose=2)
}
