% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{augmentNCRMData}
\alias{augmentNCRMData}
\title{Utility function to derive estimated probabilties of DLT toxicity band
membership for the example nCRM dataset}
\usage{
augmentNCRMData(d, doseGrid = c(3, 6, 8, 12, 16, 21, 28, 37, 46, 61, 80))
}
\arguments{
\item{d}{the runjags tibble to be augmented}

\item{doseGrid}{the doses for which p(DLT) and p(Sub-DLT or DLT) should be
calculated for each MCMC sample}
}
\value{
the augmented tibble
}
\description{
Utility function to derive estimated probabilties of DLT toxicity band
membership for the example nCRM dataset
}
