% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{augmentOCRMData}
\alias{augmentOCRMData}
\title{Utility function to derive estimated probabilties of DLT (Prob2) and DLT or
sub-DLT for the example oCRM dataset}
\usage{
augmentOCRMData(d, doseGrid = 1:10)
}
\arguments{
\item{d}{the runjags tibble to be augmented}

\item{doseGrid}{the doses for which p(DLT) and p(Sub-DLT or DLT) should be
calculated for each MCMC sample}
}
\value{
the augmented tibble
}
\description{
Utility function to derive estimated probabilties of DLT (Prob2) and DLT or
sub-DLT for the example oCRM dataset
}
