% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{augmentOQuigleyData}
\alias{augmentOQuigleyData}
\title{Utility function to derive estimated probabilties of DLT for the example
dataset based on example 1 in O'Quigley et al (1990) [pp40-41]}
\usage{
augmentOQuigleyData(d, doseGrid = 1:6)
}
\arguments{
\item{d}{the runjags tibble to be augmented}

\item{doseGrid}{the doses for which p(DLT) should be calculated for each
MCMC sample}
}
\value{
the augmented tibble
}
\description{
Utility function to derive estimated probabilties of DLT for the example
dataset based on example 1 in O'Quigley et al (1990) [pp40-41]
}
