% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createCustomDoseSummary}
\alias{createCustomDoseSummary}
\title{Create Summary Statistics from a MCMC Samples}
\usage{
createCustomDoseSummary(data, ..., widen = TRUE, group = Dose)
}
\arguments{
\item{data}{the mcmc tibble to be summarised}

\item{...}{the statistics to be produced.  See Usage notes below.}

\item{widen}{Should \code{df-col}s and \code{list} columns inthe return
value be expanded?  See Usage notes below.}

\item{group}{the name of the column in \code{data} which defines the grouping}
}
\value{
a tibble containing the requested summary statistics.  See Usage
 notes below.
}
\description{
Given a tibble equivalent of a \code{mcmc} or \code{mcmc.list} object,
creates a tibble of summary statistics by dose.
}
\section{}{
NA
}

\examples{
#Equivalent to a call to createDoseSummary:
data(oCRMPosteriorShort)
createCustomDoseSummary(
 oCRMPosteriorShort \%>\% tidy() \%>\% augmentOCRMData(),
 N=sum(!is.na(Value)),
 Q=stats::quantile(Value, q=quantiles, names=FALSE),
 Mean=mean(Value, na.rm=TRUE)
)
}
