% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createDoseSummary}
\alias{createDoseSummary}
\title{Create Summary Statistics from a MCMC Samples}
\usage{
createDoseSummary(
  data,
  group = Dose,
  quantiles = c(0.05, 0.1, 0.5, 0.9, 0.95),
  widen = TRUE
)
}
\arguments{
\item{data}{the mcmc tibble to be summarised}

\item{group}{the name of the column in \code{data} which defines the grouping}

\item{quantiles}{the quantiles to be produced.}

\item{widen}{Should \code{df-col}s and \code{list} columns in the return
value be expanded?  See Usage notes below.}
}
\value{
a tibble containing the requested summary statistics.  See Usage
 notes below.
}
\description{
Given a tibble equivalent of a \code{mcmc} or \code{mcmc.list} object,
creates a tibble of standard summary statistics by dose.
}
\section{Usage notes}{

\code{createDoseSummary} is a convenience wrapper around
\code{\link{createCustomDoseSummary}} that produces a standard summary tibble
that reports the mean and quantiles (defined by the \code{quantiles} parameter
by \code{group}.  The default action is to report the mean, 5th, 10th, 50th,
90th and 95th centiles of \code{Value} grouped by \code{Dose}.  See
 \code{\link{createCustomDoseSummary}} for more details.
}

\examples{
data(oCRMPosteriorShort)
x <- oCRMPosteriorShort \%>\% tidy() \%>\% augmentOCRMData()
createDoseSummary(x)
}
