% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{densityGraph}
\alias{densityGraph}
\title{Create a Plot of the Posterior PDFs by Dose}
\usage{
densityGraph(d, dose = Dose, value = .data$Value,
  xLabel = "p(Toxicity)", yLabel = " ", dTitle = "Dose",
  dLabels = NULL, xLimits = c(0, 1), defaultFormatting = TRUE, ...)
}
\arguments{
\item{d}{the tibble containing the posterior MCMC samples, augmented by dose,
in normal form}

\item{dose}{the column in \code{d} that defines dose}

\item{value}{the column in \code{d} that defines the probabiity of toxicity}

\item{xLabel}{the text to label the x axis}

\item{yLabel}{the text to label the y axis}

\item{dTitle}{the title of the legend describing the doses}

\item{dLabels}{the texts used to label the doses.  If \code{NA}, the raw
values of the \code{dose} column are used}

\item{xLimits}{a two element vector defining the limits of the x axis.  See
Usage nots below.}

\item{defaultFormatting}{apply default formatting to the axes and legend?
See Usage notes below.}

\item{...}{other parameters passed to \code{geom_density()}}
}
\value{
a ggplot2 object which may be further customised
}
\description{
Create a Plot of the Posterior PDFs by Dose
}
\section{Usage notes}{

The range of the x axis is limited by \code{coord_cartesian()}, so all rows
of \code{d} are used to compute the densities

If \code{default Formatting == FALSE}, the values of \code{xLabel},
\code{yLabel}, \code{dTitle} and \code{dLabels} are ignored.
}

\examples{
z <- runjagsToTibble(generateOCRMData(chainLength=1000)[["posterior"]])
z <- augmentOCRMData(z, doseGrid=4:10)
densityGraph(z, dose=Dose, value=Prob2)
}
