% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.where}
\alias{.where}
\title{local copy of tidyselect:::where, which is not excported, but needed.}
\usage{
.where(fn)
}
\arguments{
\item{fn}{the function used to filter the object}
}
\description{
local copy of tidyselect:::where, which is not excported, but needed.
}
