% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fitModel}
\alias{fitModel}
\title{Utility Function to Fit a JAGS Model}
\usage{
fitModel(
  modelString,
  observedData,
  extras = list(),
  inits = list(),
  auto = FALSE,
  ...
)
}
\arguments{
\item{modelString}{the JAGS model string}

\item{observedData}{a tibble containing the observed data}

\item{extras}{additional elements of the named list pssed to \code{runjags}}

\item{inits}{the value of the \code{inits} parameter passed to \code{runjags}}

\item{auto}{use \code{autorun.jags} or \code{run.jags}?}

\item{...}{additional arguments passed to \code{runjags}}
}
\value{
the runjags object produced by JAGS
}
\description{
The \code{observedData} tibble is converted to a list and the \code{extras}
list is then appended to it.  An additional element, "n", containing the
number of rows in \code{observdData} is added, and the combined list is
passed to \code{runjags} the \code{data} argument.
}
