% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generateNCRMData}
\alias{generateNCRMData}
\title{Generate a runjags Object Containing the Results of an nCRM Trial}
\usage{
generateNCRMData(chainLength = 60000)
}
\arguments{
\item{chainLength}{the number of samples to generate in each chain}
}
\value{
a list with two elements.  Item "patients" contains the input data;
item "posterior" contains the runjags object
}
\description{
Generate a runjags Object Containing the Results of an nCRM Trial
}
\section{Usage notes}{

The runjags object is large - about 10MB - so it is generated by this function
rather than being provided as a saved object.
}

\examples{
x <- generateNCRMData(chainLength=10000)
}
