% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generateOCRMData}
\alias{generateOCRMData}
\title{Generate patient data and the corresponding runjags Object Containing the
Results of a Fictitious oCRM Trial}
\usage{
generateOCRMData(chainLength = 120000)
}
\arguments{
\item{chainLength}{the number of samples to generate in each chain}
}
\value{
a list with two elements.  Item "patients" contains the input data;
item "posterior" contains the runjags object
}
\description{
Generate patient data and the corresponding runjags Object Containing the
Results of a Fictitious oCRM Trial
}
\section{Usage notes}{

The runjags object is large - about 1MB - so it is generated by this function
rather than being provided as a saved object.
}

\examples{
x <- generateOCRMData(chainLength=1000)
}
