% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generateOQuigleyData}
\alias{generateOQuigleyData}
\title{Generate the patient data and corresponding runjags Object for O'Quigley}
\usage{
generateOQuigleyData(chainLength = 10000)
}
\arguments{
\item{chainLength}{the number of samples to generate in each chain}
}
\value{
a list with two elements.  Item "patients" contains the input data;
item "posterior" contains the runjags object
}
\description{
Patient level data and MCMC samples for Example 1 in O'Quigley et al (1990).
}
\section{Usage notes}{

The runjags object is large - about xxxxMB - so it is generated by this function
rather than being provided as a saved object.
}

\examples{
x <- generateOQuigleyData(chainLength=1000)
}
