% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getModelString}
\alias{getModelString}
\title{Utility Function to Obtain a JAGS Model String}
\usage{
getModelString(s = c("oquigley", "ocrm", "crm"))
}
\arguments{
\item{s}{the key to identified the required model string.}
}
\value{
The requested model string
}
\description{
Obtain the model string for one of the example trials included with
crmReporter.
}
\examples{
model <- getModelString("oquigley")
}
