% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nCRMPosterior}
\alias{nCRMPosterior}
\title{The sampled posterior for the \code{nCRMPatientData} dataset}
\format{
a runjags object with 3 chains
}
\usage{
nCRMPosterior
}
\description{
A runjags object containing the output from runjags when run on the example O'Quigley dataset with a chain length of 30000
}
\keyword{datasets}
