% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nCRMPosteriorShort}
\alias{nCRMPosteriorShort}
\title{The sampled posterior for the \code{nCRMPatientData} dataset}
\format{
a runjags object with 3 chains
}
\usage{
nCRMPosteriorShort
}
\description{
A runjags object containing the output from runjags when run on the example O'Quigley dataset with a chain length of 2000
}
\keyword{datasets}
