% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oCRMPatientData}
\alias{oCRMPatientData}
\title{Patient data from a fictitious oCRM trial}
\format{
A tibble with 15 rows and 5 columns
\describe{
  \item{Cohort}{The cohort number, here, identical to the SubjectID as the model was updatd after every subject}
  \item{SubjectID}{The subject id}
  \item{Dose}{The index of the dose received by this subject.}
  \item{XHat}{The nominal dose, on the x-hat scale, received by this subject.}
  \item{Toxicity}{A binary flag indicating whether or not the subject experienced a DLT. [0=No, 1=Yes]}
}
}
\usage{
oCRMPatientData
}
\description{
Patient data from a fictitious oCRM trial
}
\keyword{datasets}
