% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oQuigleyPatientData}
\alias{oQuigleyPatientData}
\title{Documentation for datasets in /data and /data-raw}
\format{
A tibble with 25 rows and 4 columns
\describe{
  \item{Cohort}{The cohort number, here, identical to the SubjectID as the model was updatd after every subject}
  \item{SubjectID}{The subject id}
  \item{Dose}{The index of the dose received by this subject.  The actual doses used are never defined in the paper}
  \item{Toxicity}{A binary flag indicating whether or not the subject experienced a DLT. [0=No, 1=Yes]}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/2350571/}
}
\usage{
oQuigleyPatientData
}
\description{
Patient data from O'Quigley J, Pepe M & Fisher L (1990) A Practical Design for
Phase 1 Clinical Trials in Cancer.  Biometrics 46:1 33-48.  Example 1
}
\keyword{datasets}
