% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{patientSummary}
\alias{patientSummary}
\title{Summarise patient exposure by dose}
\usage{
patientSummary(
  data,
  dose = Dose,
  tox = Toxicity,
  toxLabels = NA,
  doseGrid = NULL
)
}
\arguments{
\item{data}{the mcmc tibble to be summarised}

\item{dose}{the name of the column in \code{data} which defines the dose}

\item{tox}{the name of the column in \code{data} which defines the toxicity status.  See Usage notes below.}

\item{toxLabels}{the names of the variables tahat will contain the incidence counts in the output dataset.
See Usage notes below.}

\item{doseGrid}{the values of \code{dose} for which incidence counts will be included in the output dataset.
See Usage notes below.}
}
\value{
a tibble containing one row for each value in \code{doseGrid} and columns named: the value of \code{dose},
\code{Treated}, \code{Evaluable}, and one for each element in \code{doseGrid}.  See Usage notes
below.
}
\description{
Summarise patient exposure by dose
}
\section{Usage notes}{

\code{tox} should take integer values from 0 to N, with 0 indicating no toxicity and values of 1 or more indicating
increasing values of toxicity.
If \code{toxLabels} is \code{NA}, default labelds are created dependent on the number of distinct values of
\code{tox} found in \code{data}.  If there are one or two, labels are "None" and "DLT".  If there are exactly three,
labels are "None", "SubDLT" and "DLT".  If there are four or more, labels are "None, "Cat1", ..., "CatN".
\code{doseGrid} can be used to ensure that all relevant doses appear in the output dataset, even those
 with observed zero counts.
In the output dataset, \code{Treated} counts the number of observations with non-missing values of \code{dose};
 \code{Evaluable} counts the number of observations with non-missing \code{tox}; \code{None} counts the
 number of observations wuth \code{tox} equal to 1, and the other count variables count the number of observations
  with eachand one for each of the remaining values of \code{tox}.
}

\examples{
data("oQuigleyPatientData")
patientSummary(oQuigleyPatientData, doseGrid=1:6)
}
