% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{posteriorPlot}
\alias{posteriorPlot}
\title{Plot the posterior p(Tox) by dose}
\usage{
posteriorPlot(data, dose = Dose, value = Value, recDose = NA, targetTox = NULL)
}
\arguments{
\item{data}{The tibble containing subject level dose allocations and toxicity status}

\item{dose}{the name of the column in \code{data} which defines the dose}

\item{value}{the name of the column in \code{data} which defines p(Tox) for this sample}

\item{recDose}{optional.  Identifies the currently recommended dose.  If not \code{NA}
the PDF for the given dose is drawn in  1 colour (default blue), all other doses in
another (default grey).  Otherwise, each dose is individually identified by colour}

\item{targetTox}{optional.  If not \code{NULL}, the target toxicity dose (if a scalar)
or the target toxicity band (if a vector of length 2) is highlighted in green.}
}
\description{
Plot the posterior p(Tox) by dose
}
\examples{
data("oQuigleyPosterior")
x <- oQuigleyPosterior \%>\%
       tidy() \%>\%
       augmentOQuigleyData() \%>\%
       dplyr::filter(Parameter == "Prob")
posteriorPlot(x, targetTox=c(0.2, 0.33), recDose=2)
}
