% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subjectGraph}
\alias{subjectGraph}
\title{Graph of Dose Adminsitered vs Subject Number}
\usage{
subjectGraph(d, x = SubjectID, y = Dose, cohort = Cohort,
  tox = Toxicity, toxColours = NA, toxLabels = NA,
  showCohorts = TRUE, xLabel = "Subject", yLabel = "Dose",
  mtd = NA, type = c("crm", "ocrm"), ...)
}
\arguments{
\item{d}{a tibble containing subject numbers, dose alloactions and
toxicity status.  Usually of class "ocrmSubjects" or "crmSubjects".}

\item{x}{the column in \code{d} defining the subject ID}

\item{y}{the column in \code{d} defining the dose}

\item{cohort}{the column in \code{d} defining the cohort}

\item{tox}{the column in \code{d} defining the toxicity status}

\item{toxColours}{the colours used to identify the DLT status of each
subject.  See Usage notes below.}

\item{toxLabels}{the texts used to label the various DLT statuses in the
legend.  See Usage notes below.}

\item{showCohorts}{should the graph indentify cohorts by inserting light
grey dashed lines between cohorts?}

\item{xLabel}{the text used to label the x axis}

\item{yLabel}{the text used to label the y axis}

\item{mtd}{the dose identified as the MTD at the end of the study.  Appears
in green at (maxSubject + 0.25, mtd)  where maxSubject is the maximum
subject number and mtd is the dose identified as the MTD.}

\item{type}{Optional flag to indicate the type of the trial being plotted.
Supported values are "crm" and "ocrm".  See Usage notes below.}

\item{...}{Not currently used}
}
\value{
the ggplot2 object containing the plot.
}
\description{
The graph indicates toxicity status using colur and, optionally, indicates
cohort boundaries and the MTD at the end of the trial
}
\section{Usage notes}{

If \code{toxLabels} and/or \code{toxColours} are specfied, the values
 supplied are used.  Otherwise, the value of \code{type} is used to specify
 the number of toxicity categories and the colours used to represent them.
 If \code{type == "crm"}, then  \code{toxLabels = c("None", "DLT")} and
 \code{toxColours=c("grey", "red")}.  If \code{type == "ocrm"}, then
 \code{toxLabels = c("None", "Sub-DLT", DLT")} and
 \code{toxColours=c("grey", "gold", red")}.
}

\examples{
x <- generateOCRMData(chainLength=1000)[["patients"]]
subjectGraph(x, type="ocrm")
}
