% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmPack.R
\name{summariseIncrementsRule}
\alias{summariseIncrementsRule}
\alias{summariseIncrementsRule,IncrementsRelativeDLT-method}
\alias{summariseIncrementsRule,IncrementsRelativeParts-method}
\alias{summariseIncrementsRule,IncrementsRelative-method}
\title{Create a Tibble Describing an Increments Rule}
\usage{
summariseIncrementsRule(object)

\S4method{summariseIncrementsRule}{IncrementsRelativeDLT}(object)

\S4method{summariseIncrementsRule}{IncrementsRelativeParts}(object)

\S4method{summariseIncrementsRule}{IncrementsRelative}(object)
}
\arguments{
\item{object}{An \code{Increments} object}
}
\value{
the tibble containing the description
}
\description{
Create a Tibble Describing an Increments Rule

Convert a tibble summarising an \code{Increments} rule
}
\section{Usage notes}{

The tibble contains five columns:
Type: a description of the class of the rule
LowerLimit: the lower limit of the dose range to which the \code{MaxIncrement}
applies
UpperLimit: the upper limit of the dose range to which the \code{MaxIncrement}
applies
MaxIncrement: the maximum increment for this dose range, expressed as a
decimal fraction
MaxIncrementPct: the maximum increment for this dose range, expressed as a
percentage
Lower limits are inclusive, upper limits are exclusive
}

\examples{
if (requireNamespace("crmPack")) {
  inc <- crmPack::IncrementsRelative(intervals=c(0, 20), increments=c(1, 0.33))
  summariseIncrementsRule(inc)
}
}
