% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy}
\alias{tidy}
\title{Convert MCMC Sample Chains produced by \code{runjags} to a tibble}
\usage{
tidy(
  obj,
  chain = 1,
  augment = NULL,
  indexedPattern = "([:alnum:\\\\.]*)\\\\[(\\\\d*)\\\\]",
  ...
)
}
\arguments{
\item{obj}{the object to be converted.  See Usage notes below.}

\item{chain}{the chain id}

\item{augment}{If not \code{NULL}, the function which augments the
converted \code{obj}.  The functions first argument should be the tidied version of
\code{obj}.  Other arguments can be defined in \code{...}.
with columns that are relevant to future processing.  See Usage notes below.}

\item{indexedPattern}{the regular expression used to identify indexed model
parameters.  Ensure that the pattern includes groups to return the "array"
name and the index.  The default value isn't perfect - it allows, for
example, _3a to be a valid variable name - but it sould be good enough for
most purposes.}

\item{...}{Arguments passwed to \code{augment}.}
}
\value{
the corresponding tibble.  See Usage notes below.
}
\description{
If the class of \code{obj} is \code{runjags}, the \code{mcmc} field is
extracted and processed.  If \code{runjags.list} or \code{mcmc.list},
\code{tidy} calls itself recursively on each element of
\code{obj}, \code{row_bind()}s the results together and returns the
resulting tibble.  If \code{mcmc}, \code{obj} is processed as described in
Usage notes below.
}
\section{Usage notes}{

Typically, the CRM model parameters are of little intrest in their own right.
The probabiities derived from them are more relevant.  If the JAGS code
derives these probabilities, they will be returned in the tibble as described
below.  If not, the \code{augment} argument provides a hook to allow them
to be derived.  \code{augment} is called immediately before the transformed
obj is returned, so it must process \code{obj} in its transformed state.
The transformed object is in normal form, which means that usually it must be
\code{spread()} or \code{pivot}ed \code{_wider} so that all model parameters are
available in a single row. This is slightly awkward, but takes little time.
See the vignette for an example.

Assuming that \code{augment} is \code{NULL}, the tibble returned by this
function contains the following columns:
\code{Chain}: an integer containing the index of the corresponding
\code{mcmc} object in the input \code{mcmc.list}.  If the input was an
\code{mcmc} object, \code{Chain} is \code{1} for all rows.
\code{Sample}: the sample number, from the input\code{mcmc} object.
\code{Parameter}: the name of the column in the input \code{mcmc} object to
which \code{value} relates.  If the JAGS model parameter is indexed (for
example, \code{alpha[1]}), \code{Parameter} contains the array name (here,
\code{alpha}) and \code{Index} contains the array index (here, \code{1}).
\code{Index}: if the corresponding \code{Parameter} is indexed, the value of
the index.  Otherwise, \code{NA}.
\code{Value}: The value of this \code{Parameter} drawn in this \code{Sample}
\code{Thin}: The value of the thinning parameter of the MCMC process used to generate
the samples
of this \code{Chain}.

If \code{augment} is not \code{NULL}, the return value depends on the
actions of \code{augment}.
}

\examples{
data(nCRMPosteriorShort)
tidy(nCRMPosteriorShort)
}
